web-app/requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
coloredlogs==15.0.1
decorator==5.1.1
flatbuffers==23.5.26
gitdb==4.0.10
GitPython==3.1.32
humanfriendly==10.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
lazy_loader==0.3
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
onnxruntime==1.15.1
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
platformdirs==3.9.1
pooch==1.7.0
protobuf==4.23.4
pyarrow==12.0.1
pydeck==0.8.0
Pygments==2.15.1
PyMatting==1.1.8
Pympler==1.0.1
pypng==0.20220715.0
pyshorteners==1.0.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
qrcode==7.4.2
referencing==0.30.0
rembg==2.0.50
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
scikit-image==0.21.0
scipy==1.11.1
six==1.16.0
smmap==5.0.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
tifffile==2023.7.18
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.20.0
watchdog==3.0.0
zipp==3.16.2
